requirements.txt
boto3
botocore
pypdfium2==4.30.0
faiss-cpu==1.8.0.post1
PyMuPDF==1.24.10
duckduckgo_search
langchain==0.2.16
langchain-community==0.2.17
langchain-aws==0.1.18
langchain-openai==0.1.16
openai>=1.12.0
tabula
tqdm
pillow
matplotlib
langchain_postgres
python-dotenv
pgvector==0.2.3
psycopg-binary
pinecone>=6.0.0